🛠️: refactor `Login` test to typescript
